refactor: decompose helpers.py into domain-focused modules
Split the 1146-line helpers.py into 6 focused modules following
single-responsibility and domain-cohesion principles:

- citation_metadata.py: extract and attach metadata to citations
  (add_post_citation, add_pre_citation, add_law_metadata,
   add_journal_metadata, get_year, clean_pin_cite, process_parenthetical)

- case_name.py: find case names in plain text and HTML
  (find_case_name, find_case_name_in_html, plus all helper functions)

- pin_cite.py: pin cite extraction
  (extract_pin_cite)

- court_matching.py: court code lookup from citation strings
  (get_court_by_paren)

- token_matching.py: token pattern matching utility
  (match_on_tokens, MAX_MATCH_CHARS)

- citation_filter.py: citation filtering and disambiguation
  (filter_citations, disambiguate_reporters, overlapping_citations, joke_cite)

helpers.py preserved as thin re-export module for backward compatibility.
All existing imports (from eyecite.helpers import ...) continue to work.

Verified with Regrets regression testing:
- 11 clusters: all GREEN (find-case-citation, find-law-citation,
  find-journal-citation, find-supra-citation, find-id-citation,
  clean-html, clean-whitespace, clean-underscores, resolve-citations,
  get-year-validation, get-court-by-paren)
- 2 chains: all match (full-citation-pipeline, short-cite-resolution)
- Zero drift across 5 consecutive runs
- Raw output identical to pre-refactor baseline (KEBENARAN 1)
- All fingerprints match pre-refactor baseline (KEBENARAN 2)
- Existing test suite: 20 non-hyperscan tests pass
